+++ src/Tracy/Plane.hs
@@ -7,13 +7,13 @@ import Tracy.Types
 import Tracy.Constants
 import Tracy.Util
 
-plane :: V3 Double -> V3 Double -> Material -> Object
+plane :: V3 Float -> V3 Float -> Material -> Object
 plane o n m =
     Object { _objectMaterial = m
            , _hit = hitPlane o n m
            }
 
-hitPlane :: V3 Double -> V3 Double -> Material -> Ray -> Maybe (Shade, Double)
+hitPlane :: V3 Float -> V3 Float -> Material -> Ray -> Maybe (Shade, Float)
 hitPlane o n m ray =
     let t = ((o - (ray^.origin)) `dot` n) / ((ray^.direction) `dot` n)
         hp = ray^.origin + (t *^ ray^.direction)
